+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next deploy
 
 - Documentation for installing behind a proxy
+- Ability to install behind a proxy
 - Improved docker image building time
 - Use relative paths in mount scripts
 - Run the docker image with the system user